+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.3
+
+* add query function
+    * orderByAscending
+    * orderByDescending
+    * addAscendingOrder
+    * addDescendingOrder
+    * limit
+    * skip
+* Fixed query object by objectId on Android
+
 ## 0.0.2
 
 * fixed android package error.